fix cocopack
Former-commit-id: 97037414fa6e3ebfc59571b8c0fd5542e3d430e8

+++ epbin/src/BinaryRRP.cpp
@@ -38,7 +38,7 @@ void BinaryRRP::Pack(const std::string& outfile, bool compress) const
 
 	// data sz
 	size_t data_sz = 0;
-	data_sz += sizeof(uint8_t);
+	data_sz += sizeof(int32_t);
 	for (int i = 0; i < pic_sz; ++i) {
 		data_sz += m_pics[i]->Size();
 	}
